vpn improvements and cleanups

+++ roles/wireguard_mobile/templates/container/wg-startup.sh
@@ -1,20 +1,26 @@
 #!/bin/sh
 cd /etc/wireguard/config
 
-if [ ! -f wg0.conf ]
-then
-    mkdir keys
-    mkdir clients
-
-    SERVER_ADDRESS="${VPN_NETWORK%.*}.10"
-    wg genkey | tee ./keys/server_privatekey | wg pubkey > ./keys/server_publickey
+SERVER_ADDRESS="${VPN_NETWORK%.*}.10"
 
 {% for username in vault_usernames %}
-    wg genkey | tee ./keys/client_{{username}}_privatekey | wg pubkey > ./keys/client_{{username}}_publickey
-    CLIENT_{{username}}_ADDRESS="${VPN_NETWORK%.*}.{{loop.index+10}}"
-
+CLIENT_{{username}}_ADDRESS="${VPN_NETWORK%.*}.{{loop.index+10}}"
 {% endfor %}
     
+if [ ! -f ./keys/server_privatekey || ! -f ./keys/server_publickey ]
+then
+    wg genkey | tee ./keys/server_privatekey | wg pubkey > ./keys/server_publickey
+fi
+
+{% for username in vault_usernames %}
+if [ ! -f ./keys/client_{{username}}_privatekey || ! -f ./keys/client_{{username}}_publickey ]
+then
+    wg genkey | tee ./keys/client_{{username}}_privatekey | wg pubkey > ./keys/client_{{username}}_publickey
+fi
+{% endfor %}
+    
+if [ ! -f wg0.conf ]
+then
     echo "[Interface]" > wg0.conf
     echo -n "PrivateKey = " >> wg0.conf
     cat ./keys/server_privatekey >> wg0.conf
@@ -35,9 +41,12 @@ then
 
 {% endfor %}
 
-    chmod 700 wg0.conf
+    chmod 600 wg0.conf
+fi
 
 {% for username in vault_usernames %}
+if [ ! -f ./clients/wg_{{username}}.conf ]
+then
     echo "[Interface]" > ./clients/wg_{{username}}.conf
     echo -n "Address = " >> ./clients/wg_{{username}}.conf
     echo -n $CLIENT_{{username}}_ADDRESS >> ./clients/wg_{{username}}.conf
@@ -61,19 +70,15 @@ then
     echo $PUBLIC_PORT >> ./clients/wg_{{username}}.conf
     #echo "PersistentKeepalive = 25" >> ./clients/wg_{{username}}.conf
 
-    chmod 700 ./clients/wg_{{username}}.conf
-    
+    chmod 600 ./clients/wg_{{username}}.conf
+fi
 {% endfor %}
-fi
 
-if [ -f wg0.conf ]
-then
-    wg-quick up /etc/wireguard/config/wg0.conf
-    
-    iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE
-    iptables -A FORWARD -i wg0 -j ACCEPT
-    
-    trap "echo TRAPed signal" HUP INT QUIT TERM
+wg-quick up /etc/wireguard/config/wg0.conf
 
-    tail -f /dev/null
-fi
+iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE
+iptables -A FORWARD -i wg0 -j ACCEPT
+
+trap "echo TRAPed signal" HUP INT QUIT TERM
+
+tail -f /dev/null
